+++ packages/editor-ui/src/components/Node/NodeCreator/NodeCreator.vue
@@ -1,6 +1,6 @@
 <template>
 	<div>
-		<aside :class="{'node-creator-scrim': true, expanded: !sidebarMenuCollapsed, active: showScrim}" />
+		<aside :class="{'node-creator-scrim': true, expanded: sidebarMenuCollapsed, active: showScrim}" />
 
 		<slide-transition>
 			<div