+++ packages/editor-ui/src/components/layouts/ResourcesListLayout.vue
@@ -32,8 +32,8 @@
 				<slot name="empty">
 					<n8n-action-box
 						emoji="👋"
-						:heading="$locale.baseText(currentUser.firstName ? `${resourceKey}.empty.heading` : `${resourceKey}.empty.heading.userNotSetup`, {
-							interpolate: { name: currentUser.firstName }
+						:heading="$locale.baseText(usersStore.currentUser.firstName ? `${resourceKey}.empty.heading` : `${resourceKey}.empty.heading.userNotSetup`, {
+							interpolate: { name: usersStore.currentUser.firstName }
 						})"
 						:description="$locale.baseText(`${resourceKey}.empty.description`)"
 						:buttonText="$locale.baseText(`${resourceKey}.empty.button`)"
@@ -130,6 +130,9 @@ import Vue, {PropType} from "vue";
 import {debounceHelper} from '@/components/mixins/debounce';
 import ResourceOwnershipSelect from "@/components/forms/ResourceOwnershipSelect.ee.vue";
 import ResourceFiltersDropdown from "@/components/forms/ResourceFiltersDropdown.vue";
+import { mapStores } from 'pinia';
+import { useSettingsStore } from '@/stores/settings';
+import { useUsersStore } from '@/stores/users';
 
 export interface IResource {
 	id: string;
@@ -204,20 +207,18 @@ export default mixins(
 		};
 	},
 	computed: {
-		currentUser(): IUser {
-			return this.$store.getters['users/currentUser'];
-		},
-		allUsers(): IUser[] {
-			return this.$store.getters['users/allUsers'];
-		},
+		...mapStores(
+			useSettingsStore,
+			useUsersStore,
+		),
 		subviewResources(): IResource[] {
 			if (!this.shareable) {
 				return this.resources as IResource[];
 			}
 
 			return (this.resources as IResource[]).filter((resource) => {
-				if (this.isOwnerSubview && this.$store.getters['settings/isEnterpriseFeatureEnabled'](EnterpriseEditionFeature.Sharing)) {
-					return !!(resource.ownedBy && resource.ownedBy.id === this.currentUser.id);
+				if (this.isOwnerSubview && this.settingsStore.isEnterpriseFeatureEnabled(EnterpriseEditionFeature.Sharing)) {
+					return !!(resource.ownedBy && resource.ownedBy.id === this.usersStore.currentUser?.id);
 				}
 
 				return true;
@@ -268,7 +269,7 @@ export default mixins(
 		},
 		resourcesNotOwned(): IResource[] {
 			return (this.resources as IResource[]).filter((resource) => {
-				return resource.ownedBy && resource.ownedBy.id !== this.currentUser.id;
+				return resource.ownedBy && resource.ownedBy.id !== this.usersStore.currentUser?.id;
 			});
 		},
 	},