+++ packages/nodes-base/credentials/SendGridApi.credentials.ts
@@ -14,6 +14,7 @@ export class SendGridApi implements ICredentialType {
 			displayName: 'API Key',
 			name: 'apiKey',
 			type: 'string',
+			typeOptions: { password: true },
 			default: '',
 		},
 	];