+++ packages/nodes-base/nodes/Wait/Wait.node.ts
@@ -781,7 +781,6 @@ export class Wait implements INodeType {
 		if (options.rawBody) {
 			response.binary = {
 				data: {
-					// @ts-ignore
 					data: req.rawBody.toString(BINARY_ENCODING),
 					mimeType,
 				},